+++ snuba/utils/codecs.py
